requirements.txt
streamlit>=1.28.0
boto3>=1.34.0
botocore>=1.34.0
mp-api>=0.41.0
pymatgen>=2023.10.11
qiskit>=0.45.0
qiskit-nature>=0.7.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.17.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
strands-agents>=1.17.0
strands-agents-tools>=0.2.16

# Enhanced MCP helper dependencies
ase>=3.22.1
crystal-toolkit>=2023.11.3
kaleido>=0.2.1
scipy>=1.10.0
loguru>=0.7.0

# Amazon Braket MCP Server dependencies
amazon-braket-sdk>=1.70.0
qiskit-braket-provider>=0.4.0
fastmcp>=0.5.0
pydantic>=2.0.0
pylatexenc>=2.10  # Required for Braket circuit visualization
mcp>=1.6.0  # Required for MCP server functionality